add Chunk features (#685)
* inital chunk features

* add more things

great title i know

* add some features

still no result, but we are getting there...

* make it parse

* add simple BlockPredicate

* add simple random selector

* new things

* change stuff

* base tree impl

* change things

* add more heightmaps

* fix master merge

* steps towards proper random

* fix upper level block placement logic

* Make it compile

* fix block dir serde name

* Added Tree Foliage

* more Tree work

* Implement would_survive check

I don't say its good, but it works :D

* Fixed biome placement modifier (#836)

* Implement all tree foliages

* Fix grass, flowers, seagrass, seapickles...

* Add Ores

* add Bamboo, Vine, Spring feature

Also more fixes

* Corals

* Add Tree decoration & more

Many many fixes

* fix vines

* fixed bamboo

* fix some clippy warns

* fix trees outside of chunk

* make it compile again

upsi

* fix clippy warns

* fix InsideWorldBoundsBlockPredicate

* set dirt under tress trunk

dirt, not grass

* more trees work

---------

+++ pumpkin-world/src/generation/y_offset.rs
@@ -15,14 +15,14 @@ impl YOffset {
             YOffset::BelowTop(below_top) => {
                 height as i16 - 1 + min_y as i16 - below_top.below_top as i16
             }
-            YOffset::Absolute(absolute) => absolute.absolute as i16,
+            YOffset::Absolute(absolute) => absolute.absolute,
         }
     }
 }
 
 #[derive(Deserialize)]
 pub struct Absolute {
-    absolute: u16,
+    absolute: i16,
 }
 
 #[derive(Deserialize)]
